+++ LanchesMac/ViewModels/ShoppingCartViewModel.cs
@@ -5,6 +5,6 @@ namespace LanchesMac.ViewModels
     public class ShoppingCartViewModel
     {
         public ShoppingCart ShoppingCart { get; set; }
-        public double ShoppingCartTotal { get; set; }
+        public decimal ShoppingCartTotal { get; set; }
     }
 }
